requirements/local.txt
asgiref==3.7.2
Django==4.2.8
django-ckeditor==6.7.0
django-js-asset==2.2.0
django-model-utils==4.3.1
Pillow==10.1.0
psycopg2==2.9.9
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
Unipath==1.1